refactor: 게시물 조회 테이블 ui 컴포넌트 분리
