+++ src/features/post/get-posts/ui/posts-table.tsx
@@ -1,0 +1,113 @@
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../../../../shared/ui/table/table.tsx"
+import { Edit2, MessageSquare, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
+import { Button } from "../../../../shared/ui/button/button.tsx"
+import { usePosts } from "../context.tsx"
+import { Post } from "../../../../entities/post/model.ts"
+import { highlightText } from "../../../../shared/lib/utils.tsx"
+
+type PostsTableProps = {
+  openPostDetail: (post: Post) => void
+  setSelectedPost: (post: Post) => void
+  openUserModal: (userId?: number) => void
+  setShowEditDialog: (showEditDialog: boolean) => void
+}
+
+// 게시물 테이블 렌더링
+export const PostsTable = ({ openPostDetail, setSelectedPost, openUserModal, setShowEditDialog }: PostsTableProps) => {
+  const { posts, searchOptions, loading, deletePost, setSearchOptions, updateURL } = usePosts()
+
+  // 게시물 삭제
+  const _deletePost = async (postId: number) => {
+    try {
+      await deletePost(postId)
+    } catch (error) {
+      console.error("게시물 삭제 오류:", error)
+    }
+  }
+
+  if (loading) {
+    return <div className="flex justify-center p-4">로딩 중...</div>
+  }
+
+  return (
+    <Table>
+      <TableHeader>
+        <TableRow>
+          <TableHead className="w-[50px]">ID</TableHead>
+          <TableHead>제목</TableHead>
+          <TableHead className="w-[150px]">작성자</TableHead>
+          <TableHead className="w-[150px]">반응</TableHead>
+          <TableHead className="w-[150px]">작업</TableHead>
+        </TableRow>
+      </TableHeader>
+      <TableBody>
+        {posts.map((post) => (
+          <TableRow key={post.id}>
+            <TableCell>{post.id}</TableCell>
+            <TableCell>
+              <div className="space-y-1">
+                <div>{highlightText(post.title, searchOptions.searchQuery)}</div>
+
+                <div className="flex flex-wrap gap-1">
+                  {post.tags?.map((tag) => (
+                    <span
+                      key={tag}
+                      className={`px-1 text-[9px] font-semibold rounded-[4px] cursor-pointer ${
+                        searchOptions.tag === tag
+                          ? "text-white bg-blue-500 hover:bg-blue-600"
+                          : "text-blue-800 bg-blue-100 hover:bg-blue-200"
+                      }`}
+                      onClick={() => {
+                        setSearchOptions({ tag })
+                        updateURL()
+                      }}
+                    >
+                      {tag}
+                    </span>
+                  ))}
+                </div>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div
+                className="flex items-center space-x-2 cursor-pointer"
+                onClick={() => openUserModal(post.author?.id)}
+              >
+                <img src={post.author?.image} alt={post.author?.username} className="w-8 h-8 rounded-full" />
+                <span>{post.author?.username}</span>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div className="flex items-center gap-2">
+                <ThumbsUp className="w-4 h-4" />
+                <span>{post.reactions?.likes || 0}</span>
+                <ThumbsDown className="w-4 h-4" />
+                <span>{post.reactions?.dislikes || 0}</span>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div className="flex items-center gap-2">
+                <Button variant="ghost" size="sm" onClick={() => openPostDetail(post)}>
+                  <MessageSquare className="w-4 h-4" />
+                </Button>
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={() => {
+                    setSelectedPost(post)
+                    setShowEditDialog(true)
+                  }}
+                >
+                  <Edit2 className="w-4 h-4" />
+                </Button>
+                <Button variant="ghost" size="sm" onClick={() => _deletePost(post.id)}>
+                  <Trash2 className="w-4 h-4" />
+                </Button>
+              </div>
+            </TableCell>
+          </TableRow>
+        ))}
+      </TableBody>
+    </Table>
+  )
+}
